[CodeCompletion] Move solver-based unresolved member completion to its own file
